Bug-4260: client_id is undefined when passing to axios request but is available in local storage
