+++ opencti-platform/opencti-front/src/private/Root.js
@@ -8,7 +8,6 @@ import { ConnectedThemeProvider } from '../components/AppThemeProvider';
 import Index from './Index';
 import { UserContext } from '../utils/Security';
 import AuthBoundaryComponent from './components/AuthBoundary';
-import { getAccount } from '../services/account.service';
 
 const rootPrivateQuery = graphql`
   query RootPrivateQuery {
@@ -38,10 +37,6 @@ const clearToken = () => {
   localStorage.removeItem('token');
 };
 
-const clearClientId = () => {
-  localStorage.removeItem('client_id');
-};
-
 const Root = () => (
   <AuthBoundaryComponent>
     <QueryRenderer
@@ -49,20 +44,10 @@ const Root = () => (
       variables={{}}
       render={({ props }) => {
         clearToken();
-        clearClientId();
         if (props) {
           if (props.me && props.me.access_token) {
             const token = props.me.access_token;
             localStorage.setItem('token', token);
-            getAccount().then((res) => {
-              const account = res.data;
-              if (account) {
-                const clientId = account.clients?.[0].client_id;
-                localStorage.setItem('client_id', clientId);
-              } else {
-                clearToken();
-              }
-            });
           }
           return (
             <UserContext.Provider
